requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
torch
opencv-python
numpy
twilio
matplotlib
earthengine-api
folium
streamlit-folium